requirements.txt
alembic==1.13.1
annotated-types==0.6.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
cachelib==0.13.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.7.1
Flask==3.0.2
Flask-Migrate==4.0.5
Flask-pyoidc==3.14.3
Flask-Session==0.8.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
future==1.0.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
msgspec==0.18.6
oic==1.6.1
packaging==24.0
paramiko==3.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyjwkest==1.4.2
PyJWT==2.8.0
PyNaCl==1.5.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-ldap==3.4.4
python-socketio==5.11.1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
wsproto==1.2.0
